requirements.txt
streamlit
langchain
langchain-community
langchainhub
sidrapy
matplotlib
plotly
beautifulsoup4
requests
groq
python-dotenv